add compatibility with ESP32-Arduino,
The ESP32 does not support naked ReSTART operations( START -> ReSTART), an I2C transaction must start with a START, and End with a STOP.  A ReSTART operation must be part of a START STOP sequence.  START -> ReSTART -> STOP.  So, this means that a naked ReSTART operation must be queued until a STOP is encountered.  This change cause Wire.endTransmission(false) to return an error code of 7 (continute).  Standard Arduino error checking code if(Wire.endTransmission(false)){} will always fail on ESP32-Arduino.  It must be changed to: if(Wire.endTransmission(false) !=7){} to operate properly.